+++ tests/Unit/Private/ExportJson.tests.ps1
@@ -1,0 +1,21 @@
+BeforeAll {
+    $script:dscModuleName = 'azureTUI'
+
+    Import-Module -Name $script:dscModuleName
+}
+
+AfterAll {
+    # Unload the module being tested so that it doesn't impact any other tests.
+    Get-Module -Name $script:dscModuleName -All | Remove-Module -Force
+}
+
+Describe ExportJson {
+    Context 'When calling the function with string value' {
+        It 'true equal true' {
+            InModuleScope -ModuleName $dscModuleName {
+                $true| Should -Be $true
+            }
+        }
+    }
+}
+